1. Changed UserPermission -> Role 2. Added database relations 3. Added Parcel and ShippingRequest models/tables 4. Set up UserTableSeeder + Faker Factory, Updated composer dependencies

+++ database/migrations/2013_03_02_181604_create_metadata_table.php
@@ -16,11 +16,11 @@ class CreateMetadataTable extends Migration
         Schema::create('metadata', function (Blueprint $table) {
             $table->bigIncrements('id');
             $table->string('title');
-            $table->integer('artist_id')->unsigned();
+            $table->bigInteger('artist_id')->unsigned();
             $table->foreign('artist_id')->references('id')->on('artists');
             $table->string('genre');
             $table->string('country');
-            $table->integer('release_year')->unsigned();
+            $table->smallInteger('release_year')->unsigned();
             $table->string('format');
             $table->timestamps();
         });
