+++ database/migrations/2019_03_02_181423_create_discs_table.php
@@ -15,13 +15,13 @@ class CreateDiscsTable extends Migration
     {
         Schema::create('discs', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->integer('metadata_id')->unsigned();
+            $table->bigInteger('metadata_id')->unsigned();
             $table->foreign('metadata_id')->references('id')->on('metadata');
             $table->string('condition');
             $table->string('photo_path');
             $table->integer('offer_price')->unsigned();
             $table->boolean('sold');
-            $table->integer('department_id')->unsigned();
+            $table->bigInteger('department_id')->unsigned();
             $table->foreign('department_id')->references('id')->on('departments');
             $table->timestamps();
         });
